Nikita/revamp theme (#770)
* move theme file to theme folder

* revamping theme

* simple name change

* Fxing some naming issue

* fix name issue

* Rename AppbarTheme.ts to AppBarTheme.ts.

* File name fix

* change GREY to GREY_TONE

* merge master and fix tooltiptheme and backdroptheme.
